Eric Hynds Multiselect lib updated
There was an issue with the latest versions of CDF and this JQuery
Multiselect plugin.

If you had a multiselect rendered with this plugin, it would popup to
show the options, but would never hide again.

So the core js file was updated to its latest version and the dropdown
menu works like a charm.