[CDF-896] - CDF Storage: any user is able to change the storage of another user
An impersonation guard has been implemented allowing users with admin role to work on behalf of other users.
The logic is encapsulated in the new ImpersonationHandler class which is called from StorageApi.impersonate() method.
In order to enable testability new class StorageEngineInterface is introduced and StorageEngine.setMockInstance() method
is added, which allows to provide mock for StorageEngine (the same way as this is implemented in
org.pentaho.platform.engine.security.SecurityHelper)

checkstyle

checkstyle