[SP-4887] Backport of CDF-1024 - CPF InterPluginCall changed causing CDF autoIncludes to fail. (8.2 Suite)
The second parameter of an InterPluginCall expects a valid method that exists in the class.
Originally someone passed in "" or no method, this caused the method compare to fail and gave a false report that CDA was not installed.
listQueries is a valid method for CDA, and thus will have the check work and return true.

AutoInclude logic works and was unmodified.